Revert c4570f8, c5ef703, and 582d97 due to issue #79 - back to v3 :/

+++ RobotC/Drivers/lego-ultrasound.h
@@ -5,6 +5,10 @@
  * @{
  */
 
+/*
+ * $Id: lego-ultrasound.h 133 2013-03-10 15:15:38Z xander $
+ */
+
 #ifndef __LEGOUS_H__
 #define __LEGOUS_H__
 /** \file lego-ultrasound.h
@@ -13,7 +17,7 @@
  * lego-ultrasound.h provides an API for the Lego US driver.
  *
  * License: You may use this code as you wish, provided you give credit where its due.
- * THIS CODE WILL ONLY WORK WITH ROBOTC VERSION 4.10 AND HIGHER
+ * THIS CODE WILL ONLY WORK WITH ROBOTC VERSION 3.59 AND HIGHER.
 
  *
  * Changelog:
@@ -43,7 +47,7 @@
 #define LEGOUS_CMD_RST    0x04      /*!< Command to request a warm reset */
 
 // Prototypes
-short USreadDist(tSensors link);
+int USreadDist(tSensors link);
 bool USreadDistances(tSensors link, tByteArray &distances);
 bool _USsendCmd(tSensors link, ubyte command);
 bool USsetSingleMode(tSensors link);
@@ -53,7 +57,7 @@ bool USsetEventCapture(tSensors link);
 bool USreset(tSensors link);
 
 #ifdef __HTSMUX_SUPPORT__
-short USreadDist(tMUXSensor muxsensor);
+int USreadDist(tMUXSensor muxsensor);
 
 tConfigParams LEGOUS_config = {HTSMUX_CHAN_I2C + HTSMUX_CHAN_9V + HTSMUX_CHAN_I2C_SLOW, 1, 0x02, 0x42}; /*!< Array to hold SMUX config data for sensor */
 #endif // __HTSMUX_SUPPORT__
@@ -67,25 +71,27 @@ tByteArray LEGOUS_I2CReply;
  * @return distance from the sensor or 255 if no valid range has been specified.
  */
 #ifdef __HTSMUX_SUPPORT__
-short USreadDist(tMUXSensor muxsensor) {
+int USreadDist(tMUXSensor muxsensor) {
 
   if (HTSMUXSensorTypes[muxsensor] != HTSMUXSensorCustom)
     HTSMUXconfigChannel(muxsensor, LEGOUS_config);
+
 
   if (!HTSMUXreadPort(muxsensor, LEGOUS_I2CReply, 1, 0)) {
     return 255;
   }
 
-  return (short)LEGOUS_I2CReply[0];
+  return (int)LEGOUS_I2CReply[0];
 }
 #endif
+
 
 /**
  * Get the distance values from the sensor
  * @param link the US port number
  * @return distance from the sensor or 255 if no valid range has been specified.
  */
-short USreadDist(tSensors link) {
+int USreadDist(tSensors link) {
   memset(LEGOUS_I2CRequest, 0, sizeof(tByteArray));
 
   LEGOUS_I2CRequest[0] = 2;                // Message size
@@ -97,6 +103,7 @@ short USreadDist(tSensors link) {
 
   return LEGOUS_I2CReply[0];
 }
+
 
 /**
  * Get the distance values from the sensor. The distances to the
@@ -119,6 +126,7 @@ bool USreadDistances(tSensors link, tByteArray &distances) {
   return true;
 }
 
+
 /**
  * Send a command to the US Sensor
  *
@@ -138,6 +146,7 @@ bool _USsendCmd(tSensors link, ubyte command) {
   return writeI2C(link, LEGOUS_I2CRequest);
 }
 
+
 /**
  * Configure the US sensor for Single Shot mode
  * @param link the US port number
@@ -146,6 +155,7 @@ bool _USsendCmd(tSensors link, ubyte command) {
 bool USsetSingleMode(tSensors link) {
   return _USsendCmd(link, LEGOUS_CMD_SSHOT);
 }
+
 
 /**
  * Configure the US sensor for Continuous Mode.  This is the default.
@@ -156,6 +166,7 @@ bool USsetContinuousMode(tSensors link) {
   return _USsendCmd(link, LEGOUS_CMD_CONT);
 }
 
+
 /**
  * Turn the sensor off.
  * @param link the US port number
@@ -165,6 +176,7 @@ bool USsetOff(tSensors link){
   return _USsendCmd(link, LEGOUS_CMD_OFF);
 }
 
+
 /**
  * Configure the US sensor for Event Capture mode
  * @param link the US port number
@@ -173,6 +185,7 @@ bool USsetOff(tSensors link){
 bool USsetEventCapture(tSensors link) {
   return _USsendCmd(link, LEGOUS_CMD_ECAPT);
 }
+
 
 /**
  * Request a warm reset of the sensor.
@@ -185,5 +198,8 @@ bool USreset(tSensors link) {
 
 #endif // __LEGOSNR_H__
 
+/*
+ * $Id: lego-ultrasound.h 133 2013-03-10 15:15:38Z xander $
+ */
 /* @} */
 /* @} */
